@@ -26,14 +26,17 @@
         }
     },
     "require": {
-        "php": ">=5.3.3",
-        "psr/log": "~1.0"
+        "php": "^5.6 || ^7.0",
+        "psr/log": "^1.0"
     },
     "require-dev": {
-        "monolog/monolog": "~1.11",
-        "phpunit/phpunit": "~4.3",
-        "behat/behat": "~2.5",
-        "satooshi/php-coveralls": "~0.6"
+        "behat/behat": "^2.5.5",
+        "monolog/monolog": "^1.22",
+        "phpunit/phpunit": "^5.7.20",
+        "satooshi/php-coveralls": "^1.0",
+        "symfony/console": "^2.8",
+        "symfony/dependency-injection": "^2.8",
+        "symfony/finder": "^2.8"
     },
     "suggest": {
         "psr/log-implementation": "Allows more advanced logging of the xmpp connection"
@@ -42,6 +45,12 @@
         "branch-alias": {
             "dev-master": "1.0.x-dev"
         }
+    },
+    "config": {
+        "platform": {
+            "php": "5.6"
+        },
+        "sort-packages": true
     },
     "archive": {
         "exclude": [
