@@ -37,11 +37,11 @@
 namespace Fabiang\Xmpp\EventListener\Stream;
 
 use Fabiang\Xmpp\Event\XMLEvent;
-use Fabiang\Xmpp\Exception\RuntimeException;
-use Fabiang\Xmpp\EventListener\Stream\Authentication\AuthenticationInterface;
-use Fabiang\Xmpp\Exception\Stream\AuthenticationErrorException;
 use Fabiang\Xmpp\EventListener\AbstractEventListener;
 use Fabiang\Xmpp\EventListener\BlockingEventListenerInterface;
+use Fabiang\Xmpp\EventListener\Stream\Authentication\AuthenticationInterface;
+use Fabiang\Xmpp\Exception\RuntimeException;
+use Fabiang\Xmpp\Exception\Stream\AuthenticationErrorException;
 
 /**
  * Listener
@@ -63,7 +63,7 @@ class Authentication extends AbstractEventListener implements BlockingEventListe
      *
      * @var array
      */
-    protected $mechanisms = array();
+    protected $mechanisms = [];
 
     /**
      * {@inheritDoc}
@@ -71,10 +71,10 @@ class Authentication extends AbstractEventListener implements BlockingEventListe
     public function attachEvents()
     {
         $input = $this->getConnection()->getInputStream()->getEventManager();
-        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}mechanisms', array($this, 'authenticate'));
-        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}mechanism', array($this, 'collectMechanisms'));
-        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}failure', array($this, 'failure'));
-        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}success', array($this, 'success'));
+        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}mechanisms', [$this, 'authenticate']);
+        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}mechanism', [$this, 'collectMechanisms']);
+        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}failure', [$this, 'failure']);
+        $input->attach('{urn:ietf:params:xml:ns:xmpp-sasl}success', [$this, 'success']);
     }
 
     /**
