@@ -36,10 +36,10 @@
 
 namespace Fabiang\Xmpp\EventListener\Stream;
 
+use Fabiang\Xmpp\Connection\SocketConnectionInterface;
 use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\EventListener\AbstractEventListener;
 use Fabiang\Xmpp\EventListener\BlockingEventListenerInterface;
-use Fabiang\Xmpp\Connection\SocketConnectionInterface;
 
 /**
  * Listener
@@ -62,8 +62,8 @@ class StartTls extends AbstractEventListener implements BlockingEventListenerInt
     public function attachEvents()
     {
         $input = $this->getInputEventManager();
-        $input->attach('{urn:ietf:params:xml:ns:xmpp-tls}starttls', array($this, 'starttlsEvent'));
-        $input->attach('{urn:ietf:params:xml:ns:xmpp-tls}proceed', array($this, 'proceed'));
+        $input->attach('{urn:ietf:params:xml:ns:xmpp-tls}starttls', [$this, 'starttlsEvent']);
+        $input->attach('{urn:ietf:params:xml:ns:xmpp-tls}proceed', [$this, 'proceed']);
     }
 
     /**
