@@ -36,12 +36,18 @@
 
 namespace Fabiang\Xmpp;
 
+use Fabiang\Xmpp\Connection\ConnectionInterface;
+use Fabiang\Xmpp\Event\EventManagerInterface;
+use Fabiang\Xmpp\EventListener\Logger;
+use Fabiang\Xmpp\Protocol\ImplementationInterface;
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 10:05:30.
  *
  * @coversDefaultClass Fabiang\Xmpp\Client
  */
-class ClientTest extends \PHPUnit_Framework_TestCase
+class ClientTest extends TestCase
 {
     /**
      * @var Client
@@ -76,18 +82,16 @@ class ClientTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-        $this->options = $this->getMockBuilder('Fabiang\Xmpp\Options')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $this->options = $this->createMock(Options::class);
 
-        $this->connection = $this->getMock('Fabiang\Xmpp\Connection\ConnectionInterface');
+        $this->connection = $this->createMock(ConnectionInterface::class);
 
         $this->options->expects($this->any())
             ->method('getConnection')
             ->willReturn($this->connection);
 
-        $this->eventManager = $this->getMock('Fabiang\Xmpp\Event\EventManagerInterface');
-        $this->implementation = $this->getMock('Fabiang\Xmpp\Protocol\ImplementationInterface');
+        $this->eventManager = $this->createMock(EventManagerInterface::class);
+        $this->implementation = $this->createMock(ImplementationInterface::class);
 
         $this->options->expects($this->any())
             ->method('getImplementation')
@@ -117,7 +121,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
 
         $this->implementation->expects($this->once())
             ->method('registerListener')
-            ->with($this->isInstanceOf('\Fabiang\Xmpp\EventListener\Logger'));
+            ->with($this->isInstanceOf(Logger::class));
 
         $this->implementation->expects($this->once())
             ->method('setEventManager')
@@ -150,17 +154,15 @@ class ClientTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstructorCreatingDefaults()
     {
-        $options = $this->getMockBuilder('Fabiang\Xmpp\Options')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $options = $this->createMock(Options::class);
 
         $options->expects($this->any())
             ->method('getImplementation')
             ->willReturn($this->implementation);
 
         $object = new Client($options);
-        $this->assertInstanceOf('Fabiang\Xmpp\Event\EventManagerInterface', $object->getEventManager());
-        $this->assertInstanceOf('Fabiang\Xmpp\Connection\ConnectionInterface', $object->getConnection());
+        $this->assertInstanceOf(EventManagerInterface::class, $object->getEventManager());
+        $this->assertInstanceOf(ConnectionInterface::class, $object->getConnection());
     }
 
     /**
@@ -206,7 +208,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
             ->method('send')
             ->with($this->equalTo('test'));
 
-        $message = $this->getMock('Fabiang\Xmpp\Protocol\ProtocolImplementationInterface');
+        $message = $this->createMock(Protocol\ProtocolImplementationInterface::class);
         $message->expects($this->once())
             ->method('toString')
             ->willReturn('test');
@@ -224,7 +226,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetAndGetEventManager()
     {
-        $eventManager = $this->getMock('Fabiang\Xmpp\Event\EventManagerInterface');
+        $eventManager = $this->createMock(EventManagerInterface::class);
         $this->assertSame($eventManager, $this->object->setEventManager($eventManager)->getEventManager());
     }
 
@@ -252,9 +254,12 @@ class ClientTest extends \PHPUnit_Framework_TestCase
     public function testOptionsConnection()
     {
         $options = new Options();
-        $client = $this->getMock('\Fabiang\Xmpp\Client', array('setupImplement'), array($options));
+        $client = $this->getMockBuilder(Client::class)
+            ->setMethods(array('setupImplement'))
+            ->setConstructorArgs([$options])
+            ->getMock();
 
         $optionsAssert = $client->getOptions();
-        $this->assertInstanceOf('Fabiang\Xmpp\Connection\ConnectionInterface', $optionsAssert->getConnection());
+        $this->assertInstanceOf(ConnectionInterface::class, $optionsAssert->getConnection());
     }
 }
