@@ -36,12 +36,13 @@
 
 namespace Fabiang\Xmpp\Connection;
 
-use Fabiang\Xmpp\Stream\XMLStream;
-use Fabiang\Xmpp\Stream\SocketClient;
-use Fabiang\Xmpp\EventListener\Stream\Stream;
-use Fabiang\Xmpp\Event\EventManager;
 use Fabiang\Xmpp\Event\Event;
+use Fabiang\Xmpp\Event\EventManager;
+use Fabiang\Xmpp\EventListener\Stream\Stream;
 use Fabiang\Xmpp\Options;
+use Fabiang\Xmpp\Stream\SocketClient;
+use Fabiang\Xmpp\Stream\XMLStream;
+use PHPUnit\Framework\TestCase;
 use Psr\Log\LogLevel;
 
 /**
@@ -49,7 +50,7 @@ use Psr\Log\LogLevel;
  *
  * @coversDefaultClass Fabiang\Xmpp\Connection\Socket
  */
-class SocketTest extends \PHPUnit_Framework_TestCase
+class SocketTest extends TestCase
 {
 
     /**
@@ -87,7 +88,7 @@ class SocketTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstructor()
     {
-        $mock   = $this->getMock('\Fabiang\Xmpp\Stream\SocketClient', array(), array(), '', false);
+        $mock   = $this->createMock(SocketClient::class);
         $object = new Socket($mock, '');
         $this->assertSame($mock, $object->getSocket());
     }
@@ -117,7 +118,7 @@ class SocketTest extends \PHPUnit_Framework_TestCase
         $mock   = $this->object->getSocket();
         $mock->expects($this->once())
             ->method('read')
-            ->with($this->equalTo(4096))
+            ->with($this->equalTo(65536))
             ->will($this->returnValue($return));
         $this->assertSame($return, $this->object->receive());
     }
