@@ -2,17 +2,17 @@
 
 namespace Fabiang\Xmpp\Connection;
 
-use Fabiang\Xmpp\Options;
-use Fabiang\Xmpp\Connection\Socket;
-use Fabiang\Xmpp\EventListener\Stream\Stream;
 use Fabiang\Xmpp\Event\Event;
+use Fabiang\Xmpp\EventListener\Stream\Stream;
+use Fabiang\Xmpp\Options;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-20 at 15:29:46.
  *
  * @coversDefaultClass Fabiang\Xmpp\Connection\Test
  */
-class TestTest extends \PHPUnit_Framework_TestCase
+class TestTest extends TestCase
 {
 
     /**
