@@ -36,12 +36,14 @@
 
 namespace Fabiang\Xmpp\Event;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2013-12-31 at 19:16:20.
  *
  * @coversDefaultClass Fabiang\Xmpp\Event\EventManager
  */
-class EventManagerTest extends \PHPUnit_Framework_TestCase
+class EventManagerTest extends TestCase
 {
 
     /**
