@@ -36,12 +36,14 @@
 
 namespace Fabiang\Xmpp\Event;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2013-12-31 at 20:01:01.
  *
  * @coversDefaultClass Fabiang\Xmpp\Event\Event
  */
-class EventTest extends \PHPUnit_Framework_TestCase
+class EventTest extends TestCase
 {
 
     /**
