@@ -36,10 +36,12 @@
 
 namespace Fabiang\Xmpp\Event;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-11 at 18:21:21.
  */
-class XMLEventTest extends \PHPUnit_Framework_TestCase
+class XMLEventTest extends TestCase
 {
     /**
      * @var XMLEvent
