@@ -36,17 +36,19 @@
 
 namespace Fabiang\Xmpp\EventListener;
 
-use Monolog\Logger as MonologLogger;
-use Monolog\Handler\TestHandler;
 use Fabiang\Xmpp\Event\Event;
+use Fabiang\Xmpp\Event\EventManagerInterface;
 use Fabiang\Xmpp\Options;
+use Monolog\Handler\TestHandler;
+use Monolog\Logger as MonologLogger;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-21 at 14:18:43.
  *
  * @coversDefaultClass Fabiang\Xmpp\EventListener\Logger
  */
-class LoggerTest extends \PHPUnit_Framework_TestCase
+class LoggerTest extends TestCase
 {
 
     /**
@@ -123,8 +125,8 @@ class LoggerTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetAndGetEventManager()
     {
-        $this->assertInstanceOf('\Fabiang\Xmpp\Event\EventManagerInterface', $this->object->getEventManager());
-        $eventManager = $this->getMock('\Fabiang\Xmpp\Event\EventManagerInterface');
+        $this->assertInstanceOf(EventManagerInterface::class, $this->object->getEventManager());
+        $eventManager = $this->createMock(EventManagerInterface::class);
         $this->assertSame($eventManager, $this->object->setEventManager($eventManager)->getEventManager());
     }
 }
