@@ -36,17 +36,18 @@
 
 namespace Fabiang\Xmpp\EventListener\Stream\Authentication;
 
-use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\Connection\Test;
+use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\Options;
 use Fabiang\Xmpp\Util\XML;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-27 at 12:11:12.
  *
  * @coversDefaultClass Fabiang\Xmpp\EventListener\Stream\Authentication\DigestMd5
  */
-class DigestMd5Test extends \PHPUnit_Framework_TestCase
+class DigestMd5Test extends TestCase
 {
 
     /**
