@@ -36,17 +36,18 @@
 
 namespace Fabiang\Xmpp\EventListener\Stream;
 
-use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\Connection\Test;
 use Fabiang\Xmpp\Event\EventManager;
+use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\Options;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-11 at 18:29:57.
  *
  * @coversDefaultClass Fabiang\Xmpp\EventListener\Stream\Authentication
  */
-class AuthenticationTest extends \PHPUnit_Framework_TestCase
+class AuthenticationTest extends TestCase
 {
 
     /**
@@ -276,7 +277,7 @@ class AuthenticationTest extends \PHPUnit_Framework_TestCase
         $event   = new XMLEvent;
         $event->setParameters(array($element));
 
-        $connection = $this->getMock('\Fabiang\Xmpp\Connection\Test', array());
+        $connection = $this->createMock(Test::class);
         $this->object->getOptions()->setConnection($connection);
 
         $connection->expects($this->once())
