@@ -37,15 +37,16 @@
 namespace Fabiang\Xmpp\EventListener\Stream;
 
 use Fabiang\Xmpp\Connection\Test;
-use Fabiang\Xmpp\Options;
 use Fabiang\Xmpp\Event\XMLEvent;
+use Fabiang\Xmpp\Options;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 15:11:34.
  *
  * @coversDefaultClass Fabiang\Xmpp\EventListener\Stream\Bind
  */
-class BindTest extends \PHPUnit_Framework_TestCase
+class BindTest extends TestCase
 {
 
     /**
@@ -72,6 +73,7 @@ class BindTest extends \PHPUnit_Framework_TestCase
 
         $options = new Options;
         $options->setConnection($this->connection);
+        $options->setUsername('foo/baz@test.com');
         $this->object->setOptions($options);
         $this->connection->setReady(true)->setOptions($options);
         $this->connection->connect();
@@ -131,7 +133,9 @@ class BindTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($this->object->isBlocking());
         $buffer = $this->connection->getBuffer();
         $this->assertRegExp(
-            '#<iq type="set" id="[^"]+"><bind xmlns="urn:ietf:params:xml:ns:xmpp-bind"/></iq>#',
+            '#<iq type="set" id="[^"]+">'
+            . '<bind xmlns="urn:ietf:params:xml:ns:xmpp-bind">'
+            . '<resource>baz@test.com</resource></bind></iq>#',
             $buffer[1]
         );
     }
