@@ -36,17 +36,18 @@
 
 namespace Fabiang\Xmpp\EventListener\Stream;
 
-use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\Connection\Test;
+use Fabiang\Xmpp\Event\XMLEvent;
 use Fabiang\Xmpp\Options;
 use Fabiang\Xmpp\Protocol\User\User;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-20 at 10:15:26.
  *
  * @coversDefaultClass Fabiang\Xmpp\EventListener\Stream\Roster
  */
-class RosterTest extends \PHPUnit_Framework_TestCase
+class RosterTest extends TestCase
 {
 
     /**
