@@ -37,15 +37,16 @@
 namespace Fabiang\Xmpp\EventListener\Stream;
 
 use Fabiang\Xmpp\Connection\Test;
-use Fabiang\Xmpp\Options;
 use Fabiang\Xmpp\Event\XMLEvent;
+use Fabiang\Xmpp\Options;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 14:16:00.
  *
  * @coversDefaultClass Fabiang\Xmpp\EventListener\Stream\StartTls
  */
-class StartTlsTest extends \PHPUnit_Framework_TestCase
+class StartTlsTest extends TestCase
 {
 
     /**
