@@ -37,13 +37,14 @@
 namespace Fabiang\Xmpp\Exception\Stream;
 
 use Fabiang\Xmpp\Event\XMLEvent;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 11:18:45.
  *
  * @coversDefaultClass Fabiang\Xmpp\Exception\Stream\StreamErrorException
  */
-class StreamErrorExceptionTest extends \PHPUnit_Framework_TestCase
+class StreamErrorExceptionTest extends TestCase
 {
 
     /**
