@@ -36,10 +36,12 @@
 
 namespace Fabiang\Xmpp\Exception;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 11:09:05.
  */
-class XMLParserExceptionTest extends \PHPUnit_Framework_TestCase
+class XMLParserExceptionTest extends TestCase
 {
 
     /**
