@@ -37,13 +37,14 @@
 namespace Fabiang\Xmpp;
 
 use Fabiang\Xmpp\Connection\Test;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 09:32:06.
  *
  * @coversDefaultClass Fabiang\Xmpp\Options
  */
-class OptionsTest extends \PHPUnit_Framework_TestCase
+class OptionsTest extends TestCase
 {
 
     /**
