@@ -36,15 +36,16 @@
 
 namespace Fabiang\Xmpp\Protocol;
 
-use Fabiang\Xmpp\Options;
 use Fabiang\Xmpp\Connection\Test;
+use Fabiang\Xmpp\Options;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-17 at 09:54:58.
  *
  * @coversDefaultClass Fabiang\Xmpp\Protocol\DefaultImplementation
  */
-class DefaultImplementationTest extends \PHPUnit_Framework_TestCase
+class DefaultImplementationTest extends TestCase
 {
 
     /**
