@@ -36,12 +36,14 @@
 
 namespace Fabiang\Xmpp\Protocol;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-20 at 11:12:27.
  *
  * @coversDefaultClass Fabiang\Xmpp\Protocol\Presence
  */
-class PresenceTest extends \PHPUnit_Framework_TestCase
+class PresenceTest extends TestCase
 {
 
     /**
