@@ -36,10 +36,12 @@
 
 namespace Fabiang\Xmpp\Protocol;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.1 on 2014-01-20 at 10:08:00.
  */
-class RosterTest extends \PHPUnit_Framework_TestCase
+class RosterTest extends TestCase
 {
 
     /**
