@@ -36,12 +36,14 @@
 
 namespace Fabiang\Xmpp\Stream;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2014-11-17 at 19:10:27.
  *
  * @coversDefaultClass Fabiang\Xmpp\Stream\SocketClient
  */
-class SocketClientTest extends \PHPUnit_Framework_TestCase
+class SocketClientTest extends TestCase
 {
 
     /**
