@@ -37,13 +37,14 @@
 namespace Fabiang\Xmpp\Util;
 
 use Fabiang\Xmpp\Exception\ErrorException;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2014-11-13 at 13:42:23.
  *
  * @coversDefaultClass Fabiang\Xmpp\Util\ErrorHandler
  */
-class ErrorHandlerTest extends \PHPUnit_Framework_TestCase
+class ErrorHandlerTest extends TestCase
 {
 
     /**
