@@ -36,12 +36,14 @@
 
 namespace Fabiang\Xmpp\Util;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2014-10-29 at 11:04:03.
  *
  * @coversDefaultClass Fabiang\Xmpp\Util\XML
  */
-class XMLTest extends \PHPUnit_Framework_TestCase
+class XMLTest extends TestCase
 {
 
     /**
